VG-743: mime subtype in test filename
the filename of the test medata file can contain the mime subtype, for
example: metadata0001-subtype-iso19115+xml.xml.  when it is present,
the test code uses tika to identify the mime type and compares the
expected value from the filename.  I will do some more work on the code
tomorrow.