VG-743: added pipeline test
verify AddMetaXmlTagsStep detects the right mime type

Should this test verify more than mime type, for example, the results
of the dao extract operation?

Previous versions of the testReadsVoyagerMime called the mime type
function "detect" passing a file name.  This allowed mime type
detection based on glob patterns.  The pipleline code does different.
It called calls the detect method with a buffered input stream.  So, I
changed this test code to also match the pipleline code.

The mime type of file nga-smits-ddms.xml was being detected based on
its file extension.  Since we are not currently using glob-based
detection, I renamed the file so it is not part of the junit tests.